IBX-1565: Refactor schema's types' names validation
