+++ src/Resources/config/services/schema.yml
@@ -17,7 +17,13 @@ services:
             tags:
                 - {name: 'ezplatform_graphql.schema_domain_iterator'}
 
-    EzSystems\EzPlatformGraphQL\Schema\Builder\SchemaBuilder: ~
+    EzSystems\EzPlatformGraphQL\Schema\Builder\SchemaBuilder:
+        arguments:
+            - '@Ibexa\GraphQL\Schema\Domain\NameValidator'
+        calls:
+            - method: setLogger
+              arguments:
+                  - '@logger'
 
     EzSystems\EzPlatformGraphQL\Schema\Domain\Content\Mapper\FieldDefinition\FieldDefinitionMapper:
         alias: EzSystems\EzPlatformGraphQL\Schema\Domain\Content\Mapper\FieldDefinition\DefaultFieldDefinitionMapper
@@ -83,17 +89,9 @@ services:
 
     Ibexa\GraphQL\Schema\Domain\NameValidator: ~
 
-    EzSystems\EzPlatformGraphQL\Schema\Domain\ImageVariationDomain:
-        calls:
-            - method: setLogger
-              arguments:
-                  - '@logger'
+    EzSystems\EzPlatformGraphQL\Schema\Domain\ImageVariationDomain: ~
 
-    EzSystems\EzPlatformGraphQL\Schema\Domain\Content\ContentDomainIterator:
-        calls:
-            - method: setLogger
-              arguments:
-                  - '@logger'
+    EzSystems\EzPlatformGraphQL\Schema\Domain\Content\ContentDomainIterator: ~
 
     EzSystems\EzPlatformGraphQL\Schema\Builder:
         alias: 'EzSystems\EzPlatformGraphQL\Schema\Builder\SchemaBuilder'
