+++ src/Schema/Domain/ImageVariationDomain.php
@@ -11,48 +11,25 @@ use EzSystems\EzPlatformGraphQL\Schema;
 use EzSystems\EzPlatformGraphQL\Schema\Builder;
 use EzSystems\EzPlatformGraphQL\Schema\Domain;
 use Generator;
-use Psr\Log\LoggerAwareInterface;
-use Psr\Log\LoggerAwareTrait;
-use Psr\Log\NullLogger;
-use Ibexa\GraphQL\Schema\Domain\NameValidator;
 
 /**
  * Adds configured image variations to the ImageVariationIdentifier type.
  */
-class ImageVariationDomain implements Domain\Iterator, Schema\Worker, LoggerAwareInterface
+class ImageVariationDomain implements Domain\Iterator, Schema\Worker
 {
-    use LoggerAwareTrait;
-
     const TYPE = 'ImageVariationIdentifier';
     const ARG = 'ImageVariation';
 
     /** @var \eZ\Publish\Core\MVC\ConfigResolverInterface */
     private $configResolver;
 
-    /** @var \Ibexa\GraphQL\Schema\Domain\NameValidator */
-    private $nameValidator;
-
-    public function __construct(
-        ConfigResolverInterface $configResolver,
-        NameValidator $nameValidator
-    ) {
+    public function __construct(ConfigResolverInterface $configResolver) {
         $this->configResolver = $configResolver;
-        $this->nameValidator = $nameValidator;
-        $this->logger = new NullLogger();
     }
 
     public function iterate(): Generator
     {
         foreach ($this->configResolver->getParameter('image_variations') as $identifier => $variation) {
-            if (!$this->nameValidator->isValidName($identifier)) {
-                $message = "Skipped schema generation for Image Variation with identifier '%s'. "
-                    . 'Please rename given image variation according to GraphQL specification '
-                    . '(http://spec.graphql.org/June2018/#sec-Names)';
-
-                $this->logger->warning(sprintf($message, $identifier));
-                continue;
-            }
-
             yield [self::ARG => ['identifier' => $identifier, 'variation' => $variation]];
         }
     }
